services/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
kafka-python==2.0.2
redis==4.6.0
eventlet==0.33.3
Werkzeug==2.3.7
confluent-kafka==2.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
numpy==1.24.3
flask-login==0.6.2
flask-cors==4.0.0
werkzeug==2.3.7
flask-cors
